fix: split Market row widgets into part to satisfy file-size lint (#3487)
The #3487 price-column change pushed `trade_screen.dart` to 1015
non-comment lines, tripping `repo.dart_file_non_comment_line_size`
(limit 1000). Move the four Market commodity-row widgets
(`_MarketCommodityRow`, `_MarketCommodityRowHeader`,
`_MarketCommodityRowControls`, `_StepperButton`) into a new
`trade_screen_market_row.dart` part fragment, mirroring the existing
`trade_screen_deal_book.dart` split. Behavior unchanged; library-private
classes keep using `TradeScreen` / `_MarketTabContent` statics.

Refs #3487
